pyproject.toml
[tool.poetry]
name = "papyrusai"
version = "0.1.0"
description = ""
authors = ["Wian Stipp"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.3"
fire = "^0.5.0"
requests = "^2.31.0"
Pillow = "^10.2.0"
streamlit = "^1.32.0"
configmate = {extras = ["standard"], version = "^0.1.8"}
langchain = "^0.1.11"
tqdm = "^4.66.2"
pillow-heif = "^0.16.0"
mistralai = "^1.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
